Refactor E2E tests 7/9/10
