+++ cypress/e2e/09-step_actions_from_canvas/step_actions_from_canvas.cy.js
@@ -6,10 +6,22 @@ describe('Test for Step actions from the canvas', () => {
 
         cy.openHomePage();
         cy.uploadInitialState('TimerLogCamelRoute.yaml');
+
+        cy.zoomOutXTimes(3)
     });
 
     it(' User inserts a step between two steps (+ button in between two nodes)', () => {
         cy.insertStepMiniCatalog('aggregate');
+
+        // CHECK that the step is added between two steps
+        cy.get('[data-testid="viz-step-aggregate"]').should('be.visible');
+        // CHECK that stepNodes contains of the three steps
+        cy.get('.stepNode').should('have.length', 3);
+        // CHECK that stepNodes are in the correct order
+        cy.get('.stepNode').eq(0).should('have.attr', 'data-testid', 'viz-step-timer');
+        cy.get('.stepNode').eq(1).should('have.attr', 'data-testid', 'viz-step-aggregate');
+        cy.get('.stepNode').eq(2).should('have.attr', 'data-testid', 'viz-step-log');
+
         cy.syncUpCodeChanges();
 
         // CHECK that the step is added between two steps
@@ -24,6 +36,15 @@ describe('Test for Step actions from the canvas', () => {
 
     it('In an integration with at least two steps, user deletes the first step, showing a placeholder step in its place (start-end)', () => {
         cy.deleteStep('timer');
+
+        // CHECK that the step is deleted
+        cy.get('[data-testid="viz-step-timer"]').should('not.exist');
+        // CHECK that stepNodes contains of the two steps
+        cy.get('.stepNode').should('have.length', 2);
+        // CHECK that stepNodes are in the correct order
+        cy.get('.stepNode').eq(0).should('have.attr', 'data-testid', 'viz-step-slot');
+        cy.get('.stepNode').eq(1).should('have.attr', 'data-testid', 'viz-step-log');
+
         cy.syncUpCodeChanges();
 
         // CHECK that the step is deleted
@@ -38,6 +59,16 @@ describe('Test for Step actions from the canvas', () => {
     it('In an integration with at least two steps, user deletes the first step, showing a placeholder step in its place (start-action)', () => {
         cy.insertStepMiniCatalog('arangodb');
         cy.deleteStep('timer');
+
+        // CHECK that the step is deleted
+        cy.get('[data-testid="viz-step-timer"]').should('not.exist');
+        // CHECK that stepNodes are in the correct order
+        cy.get('.stepNode').eq(0).should('have.attr', 'data-testid', 'viz-step-slot');
+        cy.get('.stepNode').eq(1).should('have.attr', 'data-testid', 'viz-step-arangodb');
+        cy.get('.stepNode').eq(2).should('have.attr', 'data-testid', 'viz-step-log');
+        // CHECK that stepNodes contains of the two steps
+        cy.get('.stepNode').should('have.length', 3);
+
         cy.syncUpCodeChanges();
 
         // CHECK that the step is deleted
@@ -53,6 +84,16 @@ describe('Test for Step actions from the canvas', () => {
     it('In an integration with at least two steps, user deletes the first step, showing a placeholder step in its place (start-action_EIP)', () => {
         cy.insertStepMiniCatalog('aggregate');
         cy.deleteStep('timer');
+
+        // CHECK that the step is deleted
+        cy.get('[data-testid="viz-step-timer"]').should('not.exist');
+        // CHECK that stepNodes contains of the three steps
+        cy.get('.stepNode').should('have.length', 3);
+        // CHECK that stepNodes are in the correct order
+        cy.get('.stepNode').eq(0).should('have.attr', 'data-testid', 'viz-step-slot');
+        cy.get('.stepNode').eq(1).should('have.attr', 'data-testid', 'viz-step-aggregate');
+        cy.get('.stepNode').eq(2).should('have.attr', 'data-testid', 'viz-step-log');
+
         cy.syncUpCodeChanges();
 
         // CHECK that the step is deleted
@@ -69,6 +110,14 @@ describe('Test for Step actions from the canvas', () => {
         cy.deleteStep('log');
         cy.appendStepMiniCatalog('timer', 'aggregate');
         cy.appendStepMiniCatalog('aggregate', 'log', 'end');
+
+        // CHECK that stepNodes are in the correct order
+        cy.get('.stepNode').eq(0).should('have.attr', 'data-testid', 'viz-step-timer');
+        cy.get('.stepNode').eq(1).should('have.attr', 'data-testid', 'viz-step-aggregate');
+        cy.get('.stepNode').eq(2).should('have.attr', 'data-testid', 'viz-step-log');
+        // CHECK that stepNodes contains of the three steps
+        cy.get('.stepNode').should('have.length', 3);
+
         cy.syncUpCodeChanges();
 
         // CHECK that stepNodes are in the correct order
@@ -85,12 +134,12 @@ describe('Test for Step actions from the canvas', () => {
         // Enable checkbox "Log mask"
         cy.get('[name="logMask"]').click();
         // CHECK that the YAML is updated logMask: 'true'
-        cy.checkCodeSpanLine('logMask', "'true'");
+        cy.checkCodeSpanLine('logMask', "true");
 
         // Disable checkbox "Log mask"
         cy.get('[name="logMask"]').click();
-        // CHECK that the YAML is updated logMask: 'false'
-        cy.checkCodeSpanLine('logMask', "'false'");
+        // CHECK that the YAML does not contain default value logMask: 'false'
+        cy.get('.code-editor').should('not.contain.text', 'logMask');
 
         // Change the value of the "groupDelay" integer field
         cy.get('[name="groupDelay"]').clear().type('15000');
