+++ cypress/e2e/09-step_actions_from_canvas/step_extension_actions_from_canvas.cy.js
@@ -6,6 +6,8 @@ describe('Test for Step extensions', () => {
 
         cy.openHomePage();
         cy.uploadInitialState('TimerLogCamelRoute.yaml');
+
+        cy.zoomOutXTimes(3)
     });
 
     it('User sees a step extension(transform)', () => {
@@ -38,58 +40,25 @@ describe('Test for Step extensions', () => {
         cy.checkCodeSpanLine('jq', "jq-test");
     });
 
-    // TODO: This test is failing due to the issue https://github.com/KaotoIO/step-extension-repository/issues/162
-    it.skip('User adds and removes a branch from a step using a step extension', () => {
+    it('User adds and removes a branch from a step using a step extension', () => {
         cy.insertStepMiniCatalog('choice');
         cy.openStepConfigurationTab('choice');
 
-        // Click on the "Add when" button, first time
-        cy.get('[data-testid="choice-add-when-button"]').click();
-        // CHECK alert message
-        cy.get('.pf-c-alert').should('contain.text', 'When: 1 added');
-        // // Close alert message
-        cy.get('.pf-c-alert__action').children('button').click();
 
-        // Click on the "Add when" button, second time
-        cy.get('[data-testid="choice-add-when-button"]').click();
-        // CHECK alert message
-        cy.get('.pf-c-alert').should('contain.text', 'When: 2 added');
-        // // Close alert message
-        cy.get('.pf-c-alert__action').children('button').click();
+        cy.addBranchChoiceExtension();
+        cy.addBranchChoiceExtension();
+        cy.addBranchChoiceExtension('otherwise');
 
-        // Click on the "Add otherwise" button
-        cy.get('[data-testid="choice-add-otherwise-button"]').click();
-        // CHECK alert message
-        cy.get('.pf-c-alert').should('contain.text', 'Otherwise added');
-        // // Close alert message
-        cy.get('.pf-c-alert__action').eq(0).children('button').click();
-
-        // Change When: 0 expression type to Jq from list
-        cy.get('[data-testid="when-0-predicate-syntax-select"]').should('be.visible').select('Jq')
-        // Type Jq expression in the first When: 0 expression field
-        cy.get('[data-testid="when-0-predicate-string-input"]').clear().type('jq-test');
-        // Type Simple expression in the second When: 1 expression field
-        cy.get('[data-testid="when-1-predicate-string-input"]').clear().type('simple-test');
-        // Click apply button
-        cy.get('[data-testid="choice-apply-button"]').click();
-
-        // CHECK that canvas has 3 new empty steps
-        cy.get('[data-testid="viz-step-slot"]').should('be.visible').and('have.length', 3);
-        // CHECK that canvas has 3 delete branch buttons
-        cy.get('[data-testid="stepNode__deleteBranch-btn"]').should('be.visible').and('have.length', 3);
+        cy.editBranchConditionChoiceExtension(0, 'jq-test', 'jq');
+        cy.editBranchConditionChoiceExtension(1, 'simple-test', 'simple');
 
         // Check that the YAML is updated
         cy.checkCodeSpanLine("jq", "jq-test");
         cy.checkCodeSpanLine("simple", "simple-test");
         cy.get('.code-editor').should('contain.text', 'otherwise')
 
-        // Delete the first When: 0 branch
-        cy.get('[data-testid="remove-branch-when-0-button"]').click();
-        cy.get('.pf-c-alert__action').children('button').click();
-
-        // Delete the otherwise branch
-        cy.get('[data-testid="remove-branch-otherwise-button"]').click();
-        cy.get('.pf-c-alert__action').children('button').click();
+        cy.deleteBranchChoiceExtension(0);
+        cy.deleteBranchChoiceExtension('otherwise');
 
         // Close Configuration tab
         cy.get('.pf-c-button.pf-m-plain').click();
