+++ cypress/e2e/10-branching_actions_from_canvas/nested_branches_step_actions_from_canvas.cy.js
@@ -25,8 +25,8 @@ describe('User completes normal actions on steps in a branch', () => {
     it('User configures a step in a branch', () => {
         cy.openStepConfigurationTab('amqp');
 
-        cy.interactWithInputObject('includeSentJMSMessageID');
-        cy.interactWithInputObject('password', 'qwerty');
+        cy.interactWithConfigInputObject('includeSentJMSMessageID');
+        cy.interactWithConfigInputObject('password', 'qwerty');
 
         // CHECK that the step yaml is updated
         cy.checkCodeSpanLine('includeSentJMSMessageID', 'true');
